MAJOR Refactor moved circuit visualization in separate module. Similar to layout visualization.
