+++ src/qce_circuit/visualization/visualize_circuit/factory_draw_components.py
@@ -27,8 +27,8 @@ from qce_circuit.structure.circuit_operations import (
     Barrier,
     VirtualPark,
 )
-from qce_circuit.visualization.intrf_draw_component import IDrawComponent
-from qce_circuit.visualization.intrf_factory_draw_components import (
+from qce_circuit.visualization.visualize_circuit.intrf_draw_component import IDrawComponent
+from qce_circuit.visualization.visualize_circuit.intrf_factory_draw_components import (
     IOperationDrawComponentFactory,
     ITransformConstructor,
 )
@@ -37,7 +37,7 @@ from qce_circuit.utilities.geometric_definitions.intrf_rectilinear_transform imp
     DynamicLength,
     DynamicPivot,
 )
-from qce_circuit.visualization.operation_components import (
+from qce_circuit.visualization.visualize_circuit.operation_components import (
     RectangleTextBlock,
     BlockRotation,
     BlockMeasure,
@@ -45,11 +45,11 @@ from qce_circuit.visualization.operation_components import (
     RotationAngle,
     SquareParkBlock,
 )
-from qce_circuit.visualization.multi_pivot_components import (
+from qce_circuit.visualization.visualize_circuit.multi_pivot_components import (
     BlockTwoQubitGate,
     BlockVerticalBarrier,
 )
-from qce_circuit.visualization.annotation_components import (
+from qce_circuit.visualization.visualize_circuit.annotation_components import (
     HorizontalVariableIndicator,
     RoundedRectangleHighlight,
 )
