+++ src/qce_circuit/visualization/visualize_circuit/icon_components.py
@@ -4,9 +4,9 @@
 from dataclasses import dataclass, field
 import numpy as np
 from matplotlib import pyplot as plt, patches as patches
-from qce_circuit.visualization.intrf_draw_component import IDrawComponent
+from qce_circuit.visualization.visualize_circuit.intrf_draw_component import IDrawComponent
 from qce_circuit.utilities.geometric_definitions import Vec2D
-from qce_circuit.visualization.style_manager import (
+from qce_circuit.visualization.visualize_circuit.style_manager import (
     StyleManager,
     IconStyleSettings,
 )
