+++ src/qce_circuit/visualization/visualize_circuit/operation_components.py
@@ -7,8 +7,8 @@ from enum import unique, Enum
 
 import numpy as np
 from matplotlib import pyplot as plt, patches as patches
-from qce_circuit.visualization.icon_components import IconMeasure
-from qce_circuit.visualization.intrf_draw_component import IDrawComponent
+from qce_circuit.visualization.visualize_circuit.icon_components import IconMeasure
+from qce_circuit.visualization.visualize_circuit.intrf_draw_component import IDrawComponent
 from qce_circuit.utilities.geometric_definitions import (
     IRectTransformComponent,
     TransformAlignment,
@@ -18,7 +18,7 @@ from qce_circuit.utilities.geometric_definitions import (
     FixedLength,
     Vec2D,
 )
-from qce_circuit.visualization.style_manager import (
+from qce_circuit.visualization.visualize_circuit.style_manager import (
     StyleManager,
     OperationStyleSettings,
     ChannelStyleSettings,
