+++ src/qce_circuit/visualization/visualize_layout/plaquette_components.py
@@ -2,11 +2,11 @@
 # Module containing rectilinear/draw components that represent layout plaquettes.
 # -------------------------------------------
 from dataclasses import dataclass, field
-from typing import List, Optional
+from typing import List
 from enum import unique, Enum, auto
 import numpy as np
 from matplotlib import pyplot as plt, patches as patches
-from qce_circuit.visualization.intrf_draw_component import IDrawComponent
+from qce_circuit.visualization.visualize_circuit.intrf_draw_component import IDrawComponent
 from qce_circuit.utilities.geometric_definitions import (
     IRectTransformComponent,
     TransformAlignment,
